requirements.txt
torch==2.4.1+cu118
torchvision==0.19.1+cu118
numpy==1.26.4
PyYAML==6.0.2
scikit-learn==1.6.1
matplotlib==3.9.2
scipy==1.14.1
pandas==2.2.3